lab5/requirements.txt
fastapi
uvicorn[standard] 
motor
pydantic
pydantic-mongo
python-dotenv 